Convert StatusBullet.js to .ts
